+++ MNE/mne/mne_hemisphere.h
@@ -177,8 +177,8 @@ public:
     fiff_int_t np;              /**< Number of source points */
     fiff_int_t ntri;            /**< Number of available triangles */
     fiff_int_t coord_frame;     /**< Coil coordinate system definition */
-    MatrixX3d rr;               /**< Source locations of available dipoles. */
-    MatrixX3d nn;               /**< Source normals of the used dipoles. */
+    MatrixX3f rr;               /**< Source locations of available dipoles. */
+    MatrixX3f nn;               /**< Source normals of the used dipoles. */
     MatrixX3i tris;             /**< Triangles */
     fiff_int_t nuse;            /**< Number of used dipoles. */
     VectorXi inuse;             /**< Used source points indicated by 1, 0 otherwise */
